hotelreservation/test/requirements.txt
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.1.1
Deprecated==1.2.13
googleapis-common-protos==1.56.4
grpcio==1.48.1
idna==3.3
packaging==21.3
protobuf==4.21.5
PyJWT==2.4.0
pymemcache==3.5.2
pymongo==4.2.0
pyparsing==3.0.9
redis==4.3.4
requests==2.28.1
six==1.16.0
urllib3==1.26.12
wrapt==1.14.1
